refactor(test): split test_proxyweb.py into topical cases/ modules
Move the 3040-line integration suite into nine topical files under test/cases/
(auth, navigation, sql_and_api, crud, settings, tables_display, query_history,
proxysql_backend, pgsql). test_proxyweb.py becomes a thin entrypoint that
waits for ProxyWeb and discovers the cases package; shared fixtures
(ProxyWebSession, constants, wait_for_proxyweb, ColoredRunner) live in
test/testlib.py. Each topical file also runs standalone.

Also:
- Restore @unittest.skipUnless(HAS_PYMYSQL, ...) decorators on
  TestProxySQL2BackendSQL, TestDigestTextDisplay, and TestZPagination, which
  were dropped by the initial slicer pass.
- Harden two pre-existing tests: test_insert_row now cleans up in a
  try/finally, and TestSettingsUIServer asserts the export response is
  successful instead of silently skipping the checks.
- Remove the orphan test/test_mdb.py (not wired into the suite).
- Update CLAUDE.md and test/README.md to describe the new layout.
