+++ test/README.md
@@ -87,34 +87,92 @@ Override credentials:
 PROXYWEB_USER=admin PROXYWEB_PASS=admin42 python3 test_proxyweb.py
 ```
 
+Run a single topical file against an already-running stack:
+
+```bash
+python3 cases/test_auth.py
+python3 cases/test_crud.py
+```
+
+Or a single test method:
+
+```bash
+python3 -m unittest cases.test_crud.TestInlinePrimaryKeyUpdate.test_update_persists_with_browser_style_pkvalues
+```
+
 ## Test coverage
+
+Tests are grouped by surface area under `cases/`. Each topical file can also be run standalone.
+
+### `cases/test_auth.py`
 
 | Class                  | What is tested                                             |
 |------------------------|------------------------------------------------------------|
 | `TestAuth`             | Login page, wrong creds, correct creds, logout, access control |
-| `TestNavigation`       | Table browsing, adhoc report, settings, config diff pages  |
-| `TestSQLExecution`     | SELECT via SQL form (including leading-whitespace regression) |
-| `TestAPIRowOperations` | Insert / update / delete a row; runtime_ table rejection; missing JSON body (400) |
-| `TestAPIConfigDiff`    | Config diff API and get_schema API                         |
-| `TestMySQLServers`     | Full CRUD on `mysql_servers` via proxyweb API              |
-| `TestQueryRules`       | Full CRUD on `mysql_query_rules` including end-to-end cycle |
-| `TestMultiServer`      | Server switching; backend isolation (mysql vs mysql2); cross-server query rule differences; independent CRUD and config diff on proxysql2 |
-| `TestProxySQL1BackendSQL` | SQL execution against ProxySQL 1 MySQL frontend            |
-| `TestProxySQL2BackendSQL` | SQL execution against ProxySQL 2 MySQL frontend            |
-| `TestConfigDiffMemoryRuntime` | Config diff between memory and runtime layers         |
-| `TestSettingsSave`     | Config save round-trip; invalid YAML rejection; missing section rejection |
-| `TestHideTables`       | hide_tables config hides/unhides tables in nav             |
-| `TestDefaultServerFallback` | Fallback when default_server name doesn't match any server |
-| `TestSettingsEditRecovery` | Settings edit page accessible without prior navigation  |
-| `TestSettingsUIServer` | UI form produces block YAML; empty server name rejected    |
-| `TestDigestTextDisplay`| digest_text truncation, whitespace, and chevron rendering  |
-| `TestZPagination`      | DataTables server-side pagination with >1000 seeded digests |
-| `TestServerSidePagination` | `/api/table_data` input validation, paging, search, sort, error handling |
 | `TestReadOnlyUser`     | Read-only user restrictions on data modification and settings |
 | `TestDefaultCredentialsHint` | Login page shows/hides default credential hint       |
 | `TestNoServersRedirect`| Empty servers config redirects admin to settings; readonly gets error |
-| `TestQueryHistory`     | Per-server query history isolation, dropdown, full page, clear |
+| `TestDefaultServerFallback` | Fallback when default_server name doesn't match any server |
+
+### `cases/test_navigation.py`
+
+| Class                  | What is tested                                             |
+|------------------------|------------------------------------------------------------|
+| `TestNavigation`       | Table browsing, adhoc report, settings, config diff pages  |
+
+### `cases/test_sql_and_api.py`
+
+| Class                  | What is tested                                             |
+|------------------------|------------------------------------------------------------|
+| `TestSQLExecution`     | SELECT via SQL form (including leading-whitespace regression) |
+| `TestAPIRowOperations` | Insert / update / delete a row; runtime_ table rejection; missing JSON body (400) |
+| `TestAPIConfigDiff`    | Config diff API and get_schema API                         |
+| `TestConfigDiffMemoryRuntime` | Config diff between memory and runtime layers         |
+
+### `cases/test_crud.py`
+
+| Class                  | What is tested                                             |
+|------------------------|------------------------------------------------------------|
+| `TestMySQLServers`     | Full CRUD on `mysql_servers` via proxyweb API              |
+| `TestQueryRules`       | Full CRUD on `mysql_query_rules` including end-to-end cycle |
+| `TestInlinePrimaryKeyUpdate` | Inline-PK UPDATE regression (ProxySQL SQLite-style `col TYPE PRIMARY KEY`) |
+
+### `cases/test_settings.py`
+
+| Class                  | What is tested                                             |
+|------------------------|------------------------------------------------------------|
+| `TestSettingsSave`     | Config save round-trip; invalid YAML rejection; missing section rejection |
+| `TestHideTables`       | hide_tables config hides/unhides tables in nav             |
+| `TestSettingsEditRecovery` | Settings edit page accessible without prior navigation  |
+| `TestSettingsUIServer` | UI form produces block YAML; empty server name rejected    |
+
+### `cases/test_tables_display.py`
+
+| Class                  | What is tested                                             |
+|------------------------|------------------------------------------------------------|
+| `TestDigestTextDisplay`| digest_text truncation, whitespace, and chevron rendering  |
+| `TestZPagination`      | DataTables server-side pagination with >1000 seeded digests |
+| `TestSmallTableClientSideMode` | Small tables use client-side DataTables mode       |
+| `TestServerSidePagination` | `/api/table_data` input validation, paging, search, sort, error handling |
 | `TestProxySQL3Autocommit` | ProxySQL 3.x autocommit compatibility                   |
+
+### `cases/test_query_history.py`
+
+| Class                  | What is tested                                             |
+|------------------------|------------------------------------------------------------|
+| `TestQueryHistory`     | Per-server query history isolation, dropdown, full page, clear |
+
+### `cases/test_proxysql_backend.py`
+
+| Class                  | What is tested                                             |
+|------------------------|------------------------------------------------------------|
+| `TestProxySQL2BackendSQL` | SQL execution against the ProxySQL MySQL frontend (:6033) |
+| `TestMultiServer`      | Server switching; backend isolation; cross-server query rule differences; independent CRUD and config diff |
+
+### `cases/test_pgsql.py`
+
+| Class                  | What is tested                                             |
+|------------------------|------------------------------------------------------------|
 | `TestPgSQLNavigation`  | PostgreSQL table browsing and navigation                   |
 | `TestPgSQLServers`     | CRUD on `pgsql_servers` via API                            |
 | `TestPgSQLUsers`       | PostgreSQL users table view and API                        |
@@ -132,7 +190,18 @@ test/
 ├── docker-compose.yml
 ├── run_tests.sh
 ├── requirements.txt          # test-only Python deps (requests)
-├── test_proxyweb.py          # integration test suite
+├── test_proxyweb.py          # thin entrypoint: wait-for-ready + discover cases/
+├── testlib.py                # shared fixtures: ProxyWebSession, ColoredRunner
+├── cases/                    # topical test modules (run standalone or via discovery)
+│   ├── test_auth.py
+│   ├── test_crud.py
+│   ├── test_navigation.py
+│   ├── test_pgsql.py
+│   ├── test_proxysql_backend.py
+│   ├── test_query_history.py
+│   ├── test_settings.py
+│   ├── test_sql_and_api.py
+│   └── test_tables_display.py
 ├── config/
 │   └── config.yml            # proxyweb config: both proxysql + proxysql2
 ├── mysql/
